Remove test configuration model
