+++ tests/Unit/Model/Test/NamedTestTest.php
@@ -5,13 +5,13 @@ declare(strict_types=1);
 namespace webignition\BasilModels\Tests\Unit\Model\Test;
 
 use webignition\BasilModels\Model\Step\StepCollection;
-use webignition\BasilModels\Model\Test\Configuration;
 use webignition\BasilModels\Model\Test\NamedTest;
 use webignition\BasilModels\Model\Test\Test;
 
 class NamedTestTest extends \PHPUnit\Framework\TestCase
 {
-    private Configuration $configuration;
+    private string $browser;
+    private string $url;
     private StepCollection $stepCollection;
     private string $path;
     private NamedTest $test;
@@ -20,16 +20,22 @@ class NamedTestTest extends \PHPUnit\Framework\TestCase
     {
         parent::setUp();
 
-        $this->configuration = new Configuration(md5((string) rand()), md5((string) rand()));
+        $this->browser = md5((string) rand());
+        $this->url = md5((string) rand());
         $this->path = md5((string) rand());
         $this->stepCollection = new StepCollection([]);
 
-        $this->test = new NamedTest(new Test($this->configuration, $this->stepCollection), $this->path);
+        $this->test = new NamedTest(new Test($this->browser, $this->url, $this->stepCollection), $this->path);
     }
 
-    public function testGetConfiguration(): void
+    public function testGetBrowser(): void
     {
-        self::assertSame($this->configuration, $this->test->getConfiguration());
+        self::assertSame($this->browser, $this->test->getBrowser());
+    }
+
+    public function testGetUrl(): void
+    {
+        self::assertSame($this->url, $this->test->getUrl());
     }
 
     public function testGetSteps(): void
